linux compile fixes
<==== Mercurial data ====>
Branch: nat_traversal

+++ common/src/utils/network/http/server/abstract_http_request_handler.cpp
@@ -97,7 +97,8 @@ namespace nx_http
             if( contentLength )
                 responseMsg.response->headers.emplace(
                     "Content-Size",
-                    QByteArray::number( contentLength.get() ) );
+                    QByteArray::number( static_cast< qulonglong >(
+                        contentLength.get() ) ) );
         }
 
         completionHandler(
